+++ changelog.md
@@ -7,3 +7,5 @@
 - Refactorized svg.rs to optimize the rendering
 - Created the latex drawer
 - Completed the documentation as of now.
+- Added a Command Line Interface
+- Pretty error printing